Improve custom target dependency on sources

+++ src/asm/lc-peak-element.c
@@ -1,13 +1,16 @@
 #include <stdio.h>
+#include <inttypes.h>
+
 
 /*
  * Driver program for lc-peak-element so only the core function must be in asm
  */
 
-int solve(int nums[], int c);
+int64_t solve(int64_t nums[], int64_t c);
 
-int main(int argc, char** argv) {
-  int i0[] = {1, 2, 3, 4, 2};
+int main(int64_t argc, char** argv) {
+  int64_t i0[] = {1, 2, 3, 4, 2};
+
   printf("%d: %d\n", 0, solve(i0, 5));
   return 0;
 }
